Ajout des checks de primary doctor pour visionner les healthfile

+++ src/main/java/ca/uqac/archicompanyproject/infra/web/healthfile/HealthFileController.java
@@ -1,5 +1,7 @@
 package ca.uqac.archicompanyproject.infra.web.healthfile;
 
+import ca.uqac.archicompanyproject.domain.caregiver.Caregiver;
+import ca.uqac.archicompanyproject.domain.caregiver.CaregiverService;
 import ca.uqac.archicompanyproject.domain.healthfile.HealthFile;
 import ca.uqac.archicompanyproject.domain.patient.Patient;
 import ca.uqac.archicompanyproject.domain.patient.PatientService;
@@ -19,9 +21,11 @@ public class HealthFileController {
 
     @PostMapping("/create/:id")
     @PreAuthorize("hasRole('CAREGIVER')")
-    public ResponseEntity<Patient> createNewHealthFileForPatientId(@RequestParam("id") Integer patientId, @RequestBody HealthFile healthFile) {
+    public ResponseEntity<Patient> createNewHealthFileForPatientId(@RequestHeader(value = "Authorization") String token, @RequestParam("id") Integer patientId, @RequestBody HealthFile healthFile) {
         try {
-            //TODO : CHECK AUTHO PAS POSSIBLE TANT QUE PRIMARY PAS IMPLEMENTE
+            if (!patientService.checkCaregiverAccessToPatient(token, patientId)) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
             Patient patient = patientService.addHealthFile(patientId, healthFile);
             return new ResponseEntity<>(patient, HttpStatus.OK);
         } catch (NotFoundException e) {
@@ -33,9 +37,11 @@ public class HealthFileController {
 
     @PutMapping("/update/:id")
     @PreAuthorize("hasRole('CAREGIVER')")
-    public ResponseEntity<HealthFile> updateHealthFileForPatientId(@RequestParam("id") Integer patientId, @RequestBody HealthFile healthFile) {
+    public ResponseEntity<HealthFile> updateHealthFileForPatientId(@RequestHeader(value = "Authorization") String token, @RequestParam("id") Integer patientId, @RequestBody HealthFile healthFile) {
         try {
-            //TODO : CHECK AUTHO PAS POSSIBLE TANT QUE PRIMARY PAS IMPLEMENTE
+            if (!patientService.checkCaregiverAccessToPatient(token, patientId)) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
             patientService.updateHealthFile(patientId, healthFile);
             return new ResponseEntity<>(healthFile, HttpStatus.OK);
         } catch (NotFoundException e) {
@@ -46,14 +52,28 @@ public class HealthFileController {
     }
 
     @GetMapping("/view/:id")
-    @PreAuthorize("hasRole('CAREGIVER')")
-    public ResponseEntity<HealthFile> viewHealthFileFromPatientId(@RequestParam("id") Integer healthFileId) {
+    @PreAuthorize("hasAnyRole('CAREGIVER','PATIENT')")
+    public ResponseEntity<HealthFile> viewHealthFileFromPatientId(@RequestHeader(value = "Authorization") String token, @RequestParam("id") Integer healthFileId) {
         try {
-            //TODO : CHECK AUTHO PAS POSSIBLE TANT QUE PRIMARY PAS IMPLEMENTE
+            //Check si authentification patient
+            Patient currentPatient = this.patientService.getPatientFromToken(token);
+            if (currentPatient.getHealthFile() == null || currentPatient.getHealthFile().getID() != healthFileId) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
             HealthFile healthFile = patientService.findHealthfileById(healthFileId);
             return new ResponseEntity<>(healthFile, HttpStatus.OK);
         } catch (NotFoundException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            try {
+                //Authentification médecin
+                if (!patientService.checkCaregiverAccessToHealthfile(token, healthFileId)) {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
+                HealthFile healthFile = patientService.findHealthfileById(healthFileId);
+                return new ResponseEntity<>(healthFile, HttpStatus.OK);
+            } catch (NotFoundException ex){
+                //vraiment pas trouvé
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -61,9 +81,11 @@ public class HealthFileController {
 
     @DeleteMapping("/delete/:id")
     @PreAuthorize("hasRole('CAREGIVER')")
-    public ResponseEntity<HealthFile> deleteHealthFileFromPatientId(@RequestParam("id") Integer healthFileId) {
+    public ResponseEntity<HealthFile> deleteHealthFileFromPatientId(@RequestHeader(value = "Authorization") String token, @RequestParam("id") Integer healthFileId) {
         try {
-            //TODO : CHECK AUTHO PAS POSSIBLE TANT QUE PRIMARY PAS IMPLEMENTE
+            if (!patientService.checkCaregiverAccessToHealthfile(token, healthFileId)) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
             patientService.deleteHealthFile(healthFileId);
             return new ResponseEntity<>(HttpStatus.OK);
         } catch (NotFoundException e) {
